requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astunparse==1.6.3
beautifulsoup4==4.13.4
cachetools==6.2.2
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.3
cycler==0.12.1
Django==5.2
et_xmlfile==2.0.0
fastapi==0.115.12
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.12.0
gast==0.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h5py==3.15.1
httplib2==0.31.0
idna==3.10
ImageIO==2.37.2
immutabledict==4.2.1
Jinja2==3.1.6
joblib==1.5.2
keras==3.12.0
keras-tuner==1.4.8
kiwisolver==1.4.9
kt-legacy==1.0.5
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.46.0
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.4
more-itertools==10.8.0
mpmath==1.3.0
namex==0.1.0
networkx==3.6.1
numba==0.63.1
numpy==2.2.5
openai-whisper==20250625
opencv-python==4.12.0.88
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.18.0
ortools==9.12.4544
packaging==25.0
pandas==2.2.3
pillow==12.0.0
proto-plus==1.26.1
protobuf==5.29.4
PuLP==3.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2025.11.3
reportlab==4.4.0
requests==2.32.3
rich==14.2.0
rsa==4.9.1
scikit-image==0.25.2
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
starlette==0.46.2
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
tifffile==2025.10.16
tiktoken==0.12.0
torch==2.9.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.2
Werkzeug==3.1.4
wheel==0.45.1
wrapt==2.0.1
